+++ includes/inter.h
@@ -4,40 +4,40 @@
 typedef struct s_candidate
 {
 	t_color	color;
-	float	dist;
+	double	dist;
 }	t_candidate;
 
 typedef struct s_hit
 {
-	float a;
-	float b;
+	double a;
+	double b;
 }				t_hit;
 
 
-float get_positive(t_hit hit);
-float just_get(t_hit hit);
+double get_positive(t_hit hit);
+double just_get(t_hit hit);
 
 t_color	inter_objects(t_cam *cam, t_vec *ray, t_scene *scene);
 t_hit	inter_sphere(t_vec *origin, t_vec *ray, t_sphere *sphere);
-float	inter_plane(t_vec *origin, t_vec *ray, t_plane *plane);
-float	inter_triang(t_vec *origin, t_vec *ray, t_triang *tri);
+double	inter_plane(t_vec *origin, t_vec *ray, t_plane *plane);
+double	inter_triang(t_vec *origin, t_vec *ray, t_triang *tri);
 
-/* t_candidate	*cand_new(t_color color, float dist); */
+/* t_candidate	*cand_new(t_color color, double dist); */
 t_color	choose_candidate(t_list **list);
 
-t_color	calc_lights(t_vec *norm, t_scene *scene, t_vec *ray, float ray_min);
-t_color	calc_lights_2s(t_vec *norm, t_scene *scene, t_vec *ray, float ray_min);
+t_color	calc_lights(t_vec *norm, t_scene *scene, t_vec *ray, double ray_min);
+t_color	calc_lights_2s(t_vec *norm, t_scene *scene, t_vec *ray, double ray_min);
 
-t_color	calc_light_matte(t_vec *n, t_light *light, t_vec *ray, float ray_len);
+t_color	calc_light_matte(t_vec *n, t_light *light, t_vec *ray, double ray_len);
 t_color	calc_light_shiny(t_sphere *sphere, t_light *light,
-			t_vec *ray, float ray_len);
+			t_vec *ray, double ray_len);
 
 typedef struct s_moller
 {
-	float	determ;
-	float	u;
-	float	v;
-	float	t;
+	double	determ;
+	double	u;
+	double	v;
+	double	t;
 	t_vec	*pvec;
 	t_vec	*tvec;
 	t_vec	*qvec;
