+++ src/scene/camera.c
@@ -1,6 +1,6 @@
 #include "minirt.h"
 
-t_cam *new_cam(t_vec *origin, t_vec *dir, float fov)
+t_cam *new_cam(t_vec *origin, t_vec *dir, double fov)
 {
 	t_cam *result;
 	result = malloc(sizeof(t_cam));
